requirements.txt
annotated-types==0.7.0
blinker==1.9.0
blis==0.7.11
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.16.0
confection==0.1.5
cryptography==44.0.2
cymem==2.0.11
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
Flask==2.3.2
google-api-core==2.24.1
google-api-python-client==2.120.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.69.0
gspread==5.12.0
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
MarkupSafe==3.0.2
murmurhash==1.0.12
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
packaging==24.2
pdfminer.six==20221105
pdfplumber==0.10.2
pillow==11.1.0
preshed==3.0.9
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
pypdfium2==4.30.1
pytesseract==0.3.10
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
setuptools==75.8.2
six==1.17.0
smart-open==6.4.0
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
thinc==8.2.5
tqdm==4.67.1
typer==0.9.4
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
wasabi==1.1.3
weasel==0.3.4
Werkzeug==3.1.3